# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/search:
    get:
      operationId: searchMembers
      tags: [members]
      summary: Search members by name or email (max 20 results)
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
  /members/{id}/memberships:
    get:
      operationId: getMemberMemberships
      tags: [members]
      summary: Get a member and all their club memberships
      parameters:
        - in: path
          name: id
          required: true
          description: Member UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Member with memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberMembershipsResponse"
        "400":
          description: Invalid member id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
        program:
          type: string
          enum: [full_time, evening]
        classYear:
          type: integer
      required:
        - id
        - fullName
        - email
        - program
        - classYear
    MembershipRecord:
      type: object
      properties:
        clubId:
          type: string
          format: uuid
        clubName:
          type: string
        clubSlug:
          type: string
        termYears:
          type: integer
        amountPaid:
          type: number
        paidAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        academicYear:
          type: string
      required:
        - clubId
        - clubName
        - clubSlug
        - termYears
        - amountPaid
        - paidAt
        - expiresAt
        - isActive
        - academicYear
    MemberMembershipsResponse:
      type: object
      properties:
        member:
          $ref: "#/components/schemas/Member"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/MembershipRecord"
      required:
        - member
        - memberships
...