# lib/api-zod/src/generated/types/membershipRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface MembershipRecord {
  clubId: string;
  clubName: string;
  clubSlug: string;
  termYears: number;
  amountPaid: number;
  paidAt: Date;
  expiresAt: Date;
  isActive: boolean;
  academicYear: string;
}